pyproject.toml
[tool.poetry]
name = "gale"
version = "0.1.0"
description = "Generates high quality articles"
authors = ["Bhaswata Choudhury <eeb20007@tezu.ac.in>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.20"
langgraph = "^0.0.48"
langchain-openai = "^0.1.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
